requirements.txt
attrs==19.3.0
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.1
json5==0.8.4
jsonschema==3.0.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
pyrsistent==0.15.6
six==1.13.0
SQLAlchemy==1.3.5
Werkzeug==0.16.0
flask-cors
flask-migrate
flask-script
flask-login
flask-wtf
flask-scss
requests
Flask-Markdown==0.3